changement de wordF et charF en vector(float) uniquement dans la classe feature

+++ Preprocessing/features.h
@@ -14,13 +14,13 @@ class Features
 public:
 
     long id;
-    std::vector<int> wordFrequency;
+    std::vector<float> wordFrequency;
     int wordCount;
-    std::vector<int> charFrequency;
+    std::vector<float> charFrequency;
     int sentenceCount;
 
     Features();
-    Features(long number, std::vector<int> wordF,int wordC,std::vector<int> charF,int sentenceC);
+    Features(long number, std::vector<float> wordF,int wordC,std::vector<float> charF,int sentenceC);
     void createcsv();
 
 };
